Make solver configurable for other strategies
